(added damage to katana and grenadelauncher so they can damage enemys made a  very unefficent shotgun i didnt know how to make one and i didnt want to watcha n tutroial so i just made one that isnt efficent but works well and also added an auto rifle which is just changeing the fire rate )

+++ Assets/scripts/BulletOwner.cs
@@ -3,7 +3,7 @@
 public class BulletOwner : MonoBehaviour
 {
     public string owner = "no owner";
-    public int normalDamage = 1;
+    public float normalDamage = 1;
     public bool isBoss = true;
-    public int criticalDamage = 1;
+    public float criticalDamage = 1;
 }
